requirements.txt
fastapi
python-socketio
uvicorn
gunicorn
pandas
jinja2
python-multipart